Refactor NDK path handling in link_to_mupdf script
Updated NDK path handling to use ANDROID_NDK_HOME or ANDROID_HOME. Improved error message for NDK not found.

+++ Builder/link_to_mupdf_1.28.3.sh
@@ -121,14 +121,16 @@ if [ "$(uname)" == "Darwin" ]; then
   FDRIOD_NDK_VERSION=$NDK_VERSION
 fi
 
-PATH1=/Users/ivanivanenko/Library/Android/sdk/ndk
-PATH2=/home/dev/Android/Sdk/ndk
+NDK="${ANDROID_NDK_HOME:-$ANDROID_HOME/ndk/$NDK_VERSION}/ndk-build"
 
-if [ ! -d "$PATH1/$NDK_VERSION" ]; then
+if [ ! -f "$NDK" ]; then
     echo "-- NDK ERROR --"
-    echo "$PATH1/$NDK_VERSION NDK NOT FOUND"
-    echo "----"
+    echo "$NDK NOT FOUND"
+    exit 1
 fi
+
+echo "Using NDK:"
+echo "$NDK"
 
 if [ "$1" == "clean_ndk" ]; then
   rm -rf $MUPDF_JAVA/obj
